llm-chatbot/requirements.txt
# --- Core ---
python-dotenv
streamlit
requests
tqdm

# --- LangChain ecosystem ---
langchain>=0.3.0,<1.0
langchain-core>=0.3.0,<1.0
langchain-community>=0.3.0,<1.0
langchain-openai>=0.2.0,<1.0
langchain-groq>=0.2.0,<1.0
langchain-huggingface>=0.1.0,<1.0

# --- Vector DB ---
faiss-cpu

# --- Embeddings ---
sentence-transformers
torch

# --- PDF processing ---
pymupdf  # used by PyMuPDFLoader
pypdf

# --- Audio / Animation ---
gtts
streamlit-lottie

# --- API clients ---
openai
groq
